Refactor root URL extracting from main interface
These should be setup helpers tailored for each user, not a part of the
scraper. All we need is some URLs to start, no need to implement a
general file importer especially when I won't know the column name
containing URLs anyway. It's the user's responsibility to give the
scraper a list of URLs, the scraper should not being doing anything
other than confirming that they are indeed proper URLs.