Add shuffling option to root URL bundling
NOTE: python docs - random - shuffle
```text
Note that even for small len(x), the total number of permutations of x can quickly grow larger than the period of most random number generators. This implies that most permutations of a long sequence can never be generated. For example, a sequence of length 2080 is the largest that can fit within the period of the Mersenne Twister random number generator.
```
> ref: [python 3.9 docs](https://docs.python.org/3.9/library/random.html)