Migrate tracing exporter from Jaeger to OpenTelemetry
OpenTelemetry format (otel.grpc/otel.http) is more flexible compared to
Jaeger format (jaeger.thrift/jaeger.grpc), allowing us to send telemetry
to more backends, including Jaeger.
And also the Jaeger format exporting library is not in a good state now.
Someone thought it's a good idea to hard code the UDP packet size to a
large value under the false assumption regarding the loopback interface
MTU.
On most Linux machines the default MTU for loopback is 65536, but on
macOS it's usually 16384. Since most of the time applications aren't
deployed on a mac, the original authors did not feel the need to
dynamically configure the max sendable UDP packet length nor the need to
do packet segmentation. And this construct got copied from Jaeger-client
to OpenTelemetry-Jaeger-exporter.
So when someone sends multiple records to a "BatchSpanProcessor",
which as the name suggests, batches requests, the message size might
exceed the loopback MTU. Without a segmentation mechanism it fails.
The owner knows this but choose not to fix it and it's too troublesome
to change kernel settings or to shrink batch size or don't batch at all
with "SimpleSpanProcessor" which increases overhead and is not what we
want.
Jaeger grpc exporter uses a deprecated protocol buffer constructor and
it emits a ton of error messages, again, kind of troublesome, especially
when Python's "warnings" library only provides coarse filtering so it's
a lot of warnings even for "one per process" mode or no warning at all
making it hard to figure out what's wrong.
Given that there's no real advantage using Jaeger format for out small
scale application, I decided to migrate to OpenTelemetry format instead.
(can still use Jaeger as our backend if needed)

reference:
- https://github.com/open-telemetry/opentelemetry-python/blob/bd8db6e2a461d896c1c0556efd0f4b569106fa96/exporter/opentelemetry-exporter-jaeger/src/opentelemetry/exporter/jaeger/send/thrift.py#L25
- https://github.com/open-telemetry/opentelemetry-python/issues/1061
- https://www.jaegertracing.io/docs/1.41/client-libraries/#emsgsize-and-udp-buffer-limits
- https://github.com/jaegertracing/jaeger-client-node/issues/124#issuecomment-324075351
- https://github.com/matrix-org/synapse/pull/13400#issue-1318857847